Took apart EventStore and SnapshotStore tests
