+++ EventSourcing.Core.Tests/EventStoreTests.cs
@@ -12,24 +12,12 @@ namespace EventSourcing.Core.Tests
   {
     protected abstract IEventStore GetEventStore();
     protected abstract IEventStore<TBaseEvent> GetEventStore<TBaseEvent>() where TBaseEvent : Event, new();
-    protected abstract ISnapshotStore GetSnapshotStore();
-    protected abstract ISnapshotStore<TBaseEvent> GetSnapshotStore<TBaseEvent>() where TBaseEvent : Event, new();
 
     [Fact]
     public async Task Can_Add_Event()
     {
       var store = GetEventStore();
       await store.AddAsync(new Event[] { new EmptyAggregate().Add(new EmptyEvent()) });
-    }
-    
-    [Fact]
-    public async Task Can_Add_Snapshot()
-    {
-      var store = GetSnapshotStore();
-      var aggregate = new SnapshotAggregate();
-      aggregate.Add(aggregate.CreateSnapshot() as SnapshotEvent);
-      var snapshot = aggregate.UncommittedEvents.First();
-      await store.AddSnapshotAsync(snapshot);
     }
 
     [Fact]
@@ -92,21 +80,6 @@ namespace EventSourcing.Core.Tests
 
       var exception = await Assert.ThrowsAnyAsync<InvalidOperationException>(
         async () => await store.AddAsync(new Event[] { e1, e2 }));
-    }
-    
-    [Fact]
-    public async Task Cannot_Add_Snapshot_With_Duplicate_AggregateId_And_Version()
-    {
-      var store = GetSnapshotStore();
-
-      var aggregate = new SnapshotAggregate();
-      aggregate.Add(aggregate.CreateSnapshot() as SnapshotEvent);
-      var snapshot = aggregate.UncommittedEvents.First();
-      
-      await store.AddSnapshotAsync(snapshot);
-
-      var exception = await Assert.ThrowsAnyAsync<ConcurrencyException>(
-        async () => await store.AddSnapshotAsync(snapshot));
     }
 
     [Fact]
@@ -193,48 +166,6 @@ namespace EventSourcing.Core.Tests
         .ToListAsync();
 
       Assert.Equal(events.Count, result.Count);
-    }
-    
-    [Fact]
-    public async Task Can_Get_Snapshot_By_AggregateId()
-    {
-      var store = GetSnapshotStore();
-
-      var aggregate = new SnapshotAggregate();
-      aggregate.Add(aggregate.CreateSnapshot() as SnapshotEvent);
-      await store.AddSnapshotAsync(aggregate.UncommittedEvents.First());
-
-      var result = await store.Snapshots
-        .Where(x => x.AggregateId == aggregate.Id)
-        .ToListAsync();
-
-      Assert.Single(result);
-    }
-    
-    [Fact]
-    public async Task Can_Get_Latest_Snapshot_By_AggregateId()
-    {
-      var store = GetSnapshotStore();
-
-      var aggregate = new SnapshotAggregate();
-      aggregate.Add(aggregate.CreateSnapshot() as SnapshotEvent);
-      aggregate.Add(new EmptyEvent());
-      aggregate.Add(aggregate.CreateSnapshot() as SnapshotEvent);
-
-      var snapshot = aggregate.UncommittedEvents[0];
-      var snapshot2 = aggregate.UncommittedEvents[2];
-      Assert.NotEqual(snapshot.AggregateVersion, snapshot2.AggregateVersion);
-      
-      await store.AddSnapshotAsync(snapshot);
-      await store.AddSnapshotAsync(snapshot2);
-
-      var result = (await store.Snapshots
-        .Where(x => x.AggregateId == aggregate.Id)
-        .OrderByDescending(x => x.AggregateVersion)
-        .ToListAsync()).FirstOrDefault();
-      
-      Assert.NotNull(result);
-      Assert.Equal(snapshot2.AggregateVersion,result.AggregateVersion);
     }
 
     [Fact]
