+++ EventSourcing.Cosmos.Tests/CosmosEventStoreTests.cs
@@ -21,12 +21,6 @@ namespace EventSourcing.Cosmos.Tests
 
     protected override IEventStore<TBaseEvent> GetEventStore<TBaseEvent>() =>
       new CosmosEventStore<TBaseEvent>(_options);
-    
-    protected override ISnapshotStore GetSnapshotStore() =>
-      new CosmosSnapshotStore(_options);
-
-    protected override ISnapshotStore<TBaseEvent> GetSnapshotStore<TBaseEvent>() =>
-      new CosmosSnapshotStore<TBaseEvent>(_options);
 
     public CosmosEventStoreTests()
     {
