+++ EventSourcing.Cosmos/CosmosEventStore.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Text.Json;
 using System.Linq;
 using System.Net;
 using System.Reflection;
